clean-up code
