+++ libs/python/generator
@@ -1,6 +1,6 @@
 #! /usr/bin/env python
-from helperGenerator import fetchEntitledServiceList
-from helperJinja2 import renderTemplateWithJson
+from helperGenerator import fetchEntitledServiceList, renderTemplateWithJson
+
 
 servicesListJsonFile = "./config/btp-base/serviceCategories.json"
 datacenterFile = "./config/btp-base/datacenters.json"
